adding Cancel application func + fix Forms size

+++ DVLD/DVLD_DataAccessLayer/Applications_Data/Applications_Data.cs
@@ -166,7 +166,7 @@ namespace DVLD_DataAccessLayer
 
         public static bool GetApplicationInfoByID( int ApplicationID,
         ref int ApplicantPersonID,ref  DateTime ApplicationDate,ref int ApplicationTypeID,
-        ref short ApplicationStatus, ref DateTime LastStatusDate, ref decimal PaidFees, ref int CreatedByUserID)
+        ref int ApplicationStatus, ref DateTime LastStatusDate, ref decimal PaidFees, ref int CreatedByUserID)
         {
 
             bool isFound = false;
@@ -195,7 +195,7 @@ namespace DVLD_DataAccessLayer
                     ApplicantPersonID = (int)reader["ApplicantPersonID"];
                     ApplicationDate = (DateTime)reader["ApplicationDate"];
                     ApplicationTypeID = (int)reader["ApplicationTypeID"];
-                    ApplicationStatus = (short)reader["ApplicationStatus"];
+                    ApplicationStatus = (int)reader["ApplicationStatus"];
                     LastStatusDate = (DateTime)reader["LastStatusDate"];
                     PaidFees = (int)reader["PaidFees"];
                     CreatedByUserID = (int)reader["CreatedByUserID"];
